yummy_api/requirements.txt
alembic==0.9.6
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
coverage==4.4.2
cryptography==1.7.2
flasgger==0.8.0
Flask==0.12.2
Flask-Cors==3.0.3
flask-heroku==0.1.9
Flask-Login==0.4.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mistune==0.8.3
nose==1.3.7
psycopg2==2.7.3.2
pyasn1==0.3.7
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
typing==3.5.3.0
urllib3==1.22
Werkzeug==0.12.2
